Add tests and create update transaction
