+++ app/services/fetch_asset_quote.rb
@@ -1,7 +1,9 @@
 require 'nokogiri'
 require 'httparty'
 
-class Crawler < ApplicationService
+class FetchAssetQuote < ApplicationService
+  SOURCE_URL_FOR_QUOTES = 'https://statusinvest.com.br/acoes/'
+
   def initialize(asset_symbol)
     @asset_symbol = asset_symbol.downcase
   end
@@ -13,7 +15,7 @@ class Crawler < ApplicationService
   private
 
   def asset_value
-    document = Nokogiri::HTML.parse(HTTParty.get("https://statusinvest.com.br/acoes/#{@asset_symbol}").body)
+    document = Nokogiri::HTML.parse(HTTParty.get("#{SOURCE_URL_FOR_QUOTES}#{@asset_symbol}").body)
       .css('div.special')
     currency = document.css('span.icon').map(&:text).first
     value = document.css('strong.value').text
