+++ spec/workers/update_asset_quote_worker_spec.rb
@@ -1,18 +1,18 @@
 require 'rails_helper'
 require 'sidekiq/testing'
 
-Sidekiq::Testing.fake!
 RSpec.describe UpdateAssetQuoteWorker, type: :worker do
+  before do
+    Sidekiq::Testing.fake!
+  end
+
   describe '.call' do
-    before do
-      @asset = Asset.create(symbol: 'PETR4', currency: 'BRL')
-    end
+    let(:asset) { Asset.create(symbol: 'PETR4', currency: 'BRL') }
     
     it 'creates a worker' do
       expect do
-        described_class.perform_async(@asset.symbol)
+        described_class.perform_async(asset.symbol)
       end.to change(described_class.jobs, :size).by(1)
-      described_class.new.perform(@asset.symbol)
     end
   end
 end
